admin/requirements.txt
flask==2.3.4
flask-cors==3.0.10
mysql-connector-python==8.1.0
Jinja2==3.1.2

